requirements.txt
pydantic==2.10.3
pydantic-settings==2.6.1
watchdog==6.0.0
openai==1.57.2
rich==13.9.4
python-dotenv==1.0.1
